fix missing bytes issue
fell into an unsigned int overflow trap - our file sizes are >> 2**31-1 (~4GB) *sigh*

+++ src/data_check.h
@@ -24,7 +24,7 @@ struct data_check_type {
     unsigned int     ntrack;
     samplerate_type  trackbitrate;
     highrestime_type time;
-    unsigned int     byte_offset;
+    uint64_t         byte_offset;
     unsigned int     vdif_frame_size; // only filled in for VDIF
     unsigned int     vdif_data_size;  // only filled in for VDIF
     threadset_t      vdif_threads; // only filled in for VDIF
